lspcon_i2c_spi: support a devpath option
Some callers may find it easier to provide the path to an I2C device
at which to communicate with the device, rather than specify the bus
number- doing so might involve trying to parse a path to extract the
number only for flashrom to do the reverse, which is error-prone and
unnecessary.

This change adds support for a `devpath` option, continuing to
allow `bus` and requiring only one of them be specified.

TEST=Verified --flash-size outputs correct values with both
     devpath=/dev/i2c-7 and bus=7, as well as noting that one is
     required if neither is specified and only one may be specified
     if both are given.

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Id2adf8a307b9205ce5e5804a6c3e22f19d0c34c9
Reviewed-on: https://review.coreboot.org/c/flashrom/+/51967
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>