spi_master: Add default write_aai function to masters that have none
write_aai is required to register a spi master, if it is not set in
spi_master struct then register_spi_master returns ERROR_FLASHROM_BUG.

Masters in this patch did not have it set in the struct, and
register_spi_master always returned an error for them. However return
value of register_spi_master was ignored, so this was hard to notice.

Next patch in the chain checks return value of register_spi_master.

BUG=b:185191942
TEST=builds and ninja test

Change-Id: I712e74e11244e1f0ab8d8e245fcd5207ce211219
Signed-off-by: Anastasia Klimchuk <aklm@chromium.org>
Reviewed-on: https://review.coreboot.org/c/flashrom/+/56102
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Nico Huber <nico.h@gmx.de>