Introduce MAINTAINERS file
The MAINTAINERS file is a list of people who take special care of
specific things and places of the tree. Also, it gives an overview
about which people are good points for contact in case of questions
appear.

In addition to that, this file is hooked up to Gerrit so that when a
patch is pushed, which touches any of these places, the related people
are added to reviewers or CC. This is done by a tool which looks up the
changed places in the MAINTAINERS file.

To clarify: This MAINTAINERS file does *not* mean that any reviews,
comments or thoughts from other people not listed there aren't welcome
anymore. We just want to make sure that patches don't get lost.

The initial file was copied from the coreboot repository. The format is
the same as the one which is used for coreboot.

Signed-off-by: Felix Singer <felixsinger@posteo.net>
Change-Id: Ic9a302055d941eeb2499a84c62c5fe1d4c9944cf
Reviewed-on: https://review.coreboot.org/c/flashrom/+/65569
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Peter Marheine <pmarheine@chromium.org>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-by: Anastasia Klimchuk <aklm@chromium.org>
Reviewed-by: Thomas Heijligen <src@posteo.de>